Move necessary storage functions to worker classes
